Seperate more tower impls into odilia-tower
