+++ odilia-tower/Cargo.toml
@@ -19,5 +19,6 @@ odilia-common.workspace = true
 pin-project = "1.1.8"
 tokio.workspace = true
 tower = { version = "0.5.2", default-features = false, features = ["util"] }
+tower-iter = { path = "../tower-iter/", version = "0.1.0" }
 tracing.workspace = true
 zbus.workspace = true
