+++ odilia-tower/src/from_state.rs
@@ -1,9 +1,9 @@
 #![allow(clippy::module_name_repetitions)]
 
-use futures::FutureExt;
-use futures_concurrency::future::Join;
+use futures::{future::ErrInto, TryFutureExt};
+use futures_concurrency::future::TryJoin;
 
-use core::{fmt::Debug, future::Future};
+use core::future::Future;
 use odilia_common::errors::OdiliaError;
 
 pub trait TryFromState<S, T>: Sized {
@@ -12,55 +12,28 @@ pub trait TryFromState<S, T>: Sized {
 	fn try_from_state(state: S, data: T) -> Self::Future;
 }
 
-impl<S, T, U1> TryFromState<S, T> for (U1,)
-where
-	U1: TryFromState<S, T>,
-	OdiliaError: From<U1::Error>,
-	T: Debug,
-{
-	type Error = OdiliaError;
-	type Future = impl Future<Output = Result<(U1,), OdiliaError>>;
-	#[tracing::instrument(skip(state))]
-	fn try_from_state(state: S, data: T) -> Self::Future {
-		(U1::try_from_state(state, data),).join().map(|(u1,)| Ok((u1?,)))
-	}
+macro_rules! impl_try_from_state {
+    ($($type:ident,)+) => {
+        impl<S, T, $($type,)+> TryFromState<S, T> for ($($type,)+)
+        where
+              $($type: TryFromState<S, T>,)+
+              $(OdiliaError: From<$type::Error>,)+
+              T: Clone,
+              S: Clone {
+            type Error = OdiliaError;
+            type Future = <($(ErrInto<$type::Future, OdiliaError>,)+) as TryJoin>::Future;
+            fn try_from_state(state: S, data: T) -> Self::Future {
+                (
+                  $($type::try_from_state(state.clone(), data.clone()).err_into(),)+
+                )
+                  .try_join()
+            }
+        }
+    }
 }
-impl<S, T, U1, U2> TryFromState<S, T> for (U1, U2)
-where
-	U1: TryFromState<S, T>,
-	U2: TryFromState<S, T>,
-	OdiliaError: From<U1::Error> + From<U2::Error>,
-	S: Clone,
-	T: Clone + Debug,
-{
-	type Error = OdiliaError;
-	type Future = impl Future<Output = Result<(U1, U2), OdiliaError>>;
-	#[tracing::instrument(skip(state))]
-	fn try_from_state(state: S, data: T) -> Self::Future {
-		(U1::try_from_state(state.clone(), data.clone()), U2::try_from_state(state, data))
-			.join()
-			.map(|(u1, u2)| Ok((u1?, u2?)))
-	}
-}
-impl<S, T, U1, U2, U3> TryFromState<S, T> for (U1, U2, U3)
-where
-	U1: TryFromState<S, T>,
-	U2: TryFromState<S, T>,
-	U3: TryFromState<S, T>,
-	OdiliaError: From<U1::Error> + From<U2::Error> + From<U3::Error>,
-	S: Clone,
-	T: Clone + Debug,
-{
-	type Error = OdiliaError;
-	type Future = impl Future<Output = Result<(U1, U2, U3), OdiliaError>>;
-	#[tracing::instrument(skip(state))]
-	fn try_from_state(state: S, data: T) -> Self::Future {
-		(
-			U1::try_from_state(state.clone(), data.clone()),
-			U2::try_from_state(state.clone(), data.clone()),
-			U3::try_from_state(state, data),
-		)
-			.join()
-			.map(|(u1, u2, u3)| Ok((u1?, u2?, u3?)))
-	}
-}
+
+impl_try_from_state!(U1,);
+impl_try_from_state!(U1, U2,);
+impl_try_from_state!(U1, U2, U3,);
+impl_try_from_state!(U1, U2, U3, U4,);
+impl_try_from_state!(U1, U2, U3, U4, U5,);
