+++ odilia/src/tower/mod.rs
@@ -1,8 +1,8 @@
-pub mod async_try;
+pub use odilia_tower::async_try;
 pub mod choice;
 pub mod extractors;
 pub use extractors::*;
-pub mod from_state;
+pub use odilia_tower::from_state;
 pub mod handler;
 pub use tower_iter::iter_svc;
 pub mod service_ext;
@@ -10,8 +10,8 @@ pub use tower_iter::service_set;
 pub mod state_changed;
 pub mod state_svc;
 pub mod sync_try;
-pub mod unwrap_svc;
 pub use handler::Handler;
+pub use odilia_tower::unwrap_svc;
 pub use service_ext::ServiceExt;
 
 pub mod handlers;
